feat: update Profile and UserAggregate to use Long for userId; refactor related service methods and tests

+++ api/src/main/java/dev/kalbarczyk/api/core/composite/user/UserAggregate.java
@@ -1,7 +1,7 @@
 package dev.kalbarczyk.api.core.composite.user;
 
 public record UserAggregate(
-        int userId,
+        Long userId,
         String username,
         String email,
         String displayName,
