+++ api/src/main/java/dev/kalbarczyk/api/core/profile/Profile.java
@@ -1,5 +1,5 @@
 package dev.kalbarczyk.api.core.profile;
 
 
-public record Profile(int userId, String displayName, String avatarUrl, String bio, String location,String createdAt,String updatedAt) {
+public record Profile(Long userId, String displayName, String avatarUrl, String bio, String location,String createdAt,String updatedAt) {
 }
